requirements.txt
# ADReSS – Automated Disaster Response Scenario Generation and Simulation

# Dashboard core
streamlit==1.55.0
altair==6.0.0
folium==0.20.0
streamlit-folium==0.26.2

# Data I/O
pandas==2.3.3
numpy==2.4.2
openpyxl==3.1.5
PyYAML==6.0.3

# API / distance
requests==2.32.5
haversine==2.9.0

# Statistics (used in Analytics tab)
scipy==1.17.1
statsmodels==0.14.6
scikit-posthocs==0.12.0
pingouin==0.6.0

# Simulation / RL environment (MCIEnvironment_gymnasium.py)
gymnasium==1.2.3

# 3D plot for results comparison
plotly==6.6.0

# Experiment pipeline (experiment_1/)
geopandas>=1.0.0
shapely>=2.0.0
matplotlib>=3.8.0
